WIP: began moving parse functions into type modules
