+++ src/parser/types/index/mod.rs
@@ -1,4 +1,5 @@
-use crate::parser::Sql;
+use crate::parser::{Rule, Sql};
+use pest::iterators::Pair;
 
 #[derive(Debug, Clone)]
 pub struct Index {
@@ -14,6 +15,23 @@ impl Index {
             columns,
             unique,
         }
+    }
+}
+
+impl From<Pair<'_, Rule>> for Index {
+    fn from(pair: Pair<'_, Rule>) -> Self {
+        let mut inner = pair.into_inner();
+        let index_type = inner.next().unwrap().as_str();
+        let name = inner
+            .next()
+            .map(|p| p.as_str().trim_matches('`').to_string())
+            .unwrap_or_else(|| format!("index_{}", uuid::Uuid::new_v4()));
+        let columns: Vec<String> = inner
+            .map(|col| col.as_str().trim_matches('`').to_string())
+            .collect();
+        let unique = index_type.contains("UNIQUE") || index_type == "PRIMARY KEY";
+
+        Index::new(name, columns, unique)
     }
 }
